New Presets: REFACTORED KEYS  + integrated with config boxes



    integrated config boxes into presets
    moved ~100 keys from print to tool-print and other keys shuffled
    exception handling reduced to top-level only

Keys moved from print to tool-print: bottom_solid_layers, bottom_solid_min_thickness, bridge_acceleration, bridge_flow_ratio, bridge_speed, default_acceleration, dont_support_bridges, enable_dynamic_overhang_speeds, external_perimeter_acceleration, external_perimeter_extrusion_width, external_perimeter_speed, extrusion_width, first_layer_acceleration, first_layer_acceleration_over_raft, first_layer_extrusion_width, first_layer_infill_speed, first_layer_speed, first_layer_speed_over_raft, gap_fill_enabled, gap_fill_speed, infill_acceleration, infill_anchor, infill_anchor_max, infill_every_layers, infill_extrusion_width, infill_overlap, infill_speed, max_print_speed, max_volumetric_extrusion_rate_slope_negative, max_volumetric_extrusion_rate_slope_positive, max_volumetric_speed, min_bead_width, min_feature_size, only_retract_when_crossing_perimeters, over_bridge_speed, overhang_speed_0, overhang_speed_1, overhang_speed_2, overhang_speed_3, perimeter_acceleration, perimeter_extrusion_width, perimeter_speed, perimeters, raft_contact_distance, raft_expansion, raft_first_layer_density, raft_first_layer_expansion, small_perimeter_speed, solid_infill_acceleration, solid_infill_extrusion_width, solid_infill_speed, support_material_bottom_contact_distance, support_material_bottom_interface_layers, support_material_contact_distance, support_material_extrusion_width, support_material_interface_contact_loops, support_material_interface_layers, support_material_interface_pattern, support_material_interface_spacing, support_material_interface_speed, support_material_pattern, support_material_spacing, support_material_speed, support_material_with_sheath, support_material_xy_spacing, support_tree_angle, support_tree_angle_slow, support_tree_branch_diameter, support_tree_branch_diameter_angle, support_tree_branch_diameter_double_wall, support_tree_branch_distance, support_tree_tip_diameter, support_tree_top_rate, thick_bridges, top_infill_extrusion_width, top_solid_infill_acceleration, top_solid_infill_speed, top_solid_layers, top_solid_min_thickness, travel_acceleration, travel_speed, travel_speed_z, wipe_tower_extra_flow, wipe_tower_extra_spacing, brim_separation, elefant_foot_compensation, fill_density, fill_pattern , gcode_resolution, overhangs, seam_position, support_material, support_material_angle, support_material_auto, support_material_buildplate_only, support_material_closing_radius, support_material_enforce_layers, support_material_extruder, support_material_interface_extruder, support_material_style, support_material_synchronize_layers, support_material_threshold

Keys moved to printer: extruder_clearance_height, extruder_clearance_radius, extruder_offset (type changed Vec2d -> std::vector)

Key moved to print: single_extruder_multi_material_priming

+++ src/slic3r-shared/src/Slic3r/App/Scene/GeometryDataFactory.cpp
@@ -125,7 +125,7 @@ void GeometryDataFactory::create_circle()
     Render::GeometryBuilder<Render::VertexP3> builder;
     for (uint8_t i = 0; i < CIRCLE_RES; ++i) {
         float angle = i * CIRCLE_ANGLE_STEP;
-        builder.add_vertex({{ 0.5f * cos(angle), 0.5f * sin(angle), 0.0f }});
+        builder.add_vertex({{ 0.5f * cosf(angle), 0.5f * sinf(angle), 0.0f }});
     }
     builder
         .add_draw_command({ Render::PrimitiveType::LineLoop, 0, CIRCLE_RES,  Render::Material{} });
@@ -138,21 +138,21 @@ void GeometryDataFactory::create_graded_circle()
     for (uint8_t i = 0; i < CIRCLE_RES; ++i) {
         float angle_i = i * CIRCLE_ANGLE_STEP;
         float angle_j = (i + 1) * CIRCLE_ANGLE_STEP;
-        builder.add_vertex({ { 0.5f * cos(angle_i), 0.5f * sin(angle_i), 0.0f } });
-        builder.add_vertex({ { 0.5f * cos(angle_j), 0.5f * sin(angle_j), 0.0f } });
+        builder.add_vertex({ { 0.5f * cosf(angle_i), 0.5f * sinf(angle_i), 0.0f } });
+        builder.add_vertex({ { 0.5f * cosf(angle_j), 0.5f * sinf(angle_j), 0.0f } });
     }
     float grade_step = float(TWO_PI) / CIRCLE_FINE_GRADE_SECONDARY_STEPS;
     for (uint8_t i = 0; i < CIRCLE_FINE_GRADE_SECONDARY_STEPS; ++i) {
         float angle_i = i * grade_step;
         float out_radius = (i % 2 == 0) ? 0.5f * CIRCLE_FINE_GRADE_PRIMARY_OUT_RADIUS : 0.5f * CIRCLE_FINE_GRADE_SECONDARY_OUT_RADIUS;
-        builder.add_vertex({ { 0.5f * cos(angle_i), 0.5f * sin(angle_i), 0.0f } });
-        builder.add_vertex({ { out_radius * cos(angle_i), out_radius * sin(angle_i), 0.0f } });
+        builder.add_vertex({ { 0.5f * cosf(angle_i), 0.5f * sinf(angle_i), 0.0f } });
+        builder.add_vertex({ { out_radius * cosf(angle_i), out_radius * sinf(angle_i), 0.0f } });
     }
     grade_step = float(TWO_PI) / CIRCLE_COARSE_GRADE_STEPS;
     for (uint8_t i = 0; i < CIRCLE_COARSE_GRADE_STEPS; ++i) {
         float angle_i = i * grade_step;
-        builder.add_vertex({ { 0.5f * CIRCLE_COARSE_GRADE_IN_RADIUS * cos(angle_i), 0.5f * CIRCLE_COARSE_GRADE_IN_RADIUS * sin(angle_i), 0.0f } });
-        builder.add_vertex({ { 0.5f * CIRCLE_COARSE_GRADE_OUT_RADIUS * cos(angle_i), 0.5f * CIRCLE_COARSE_GRADE_OUT_RADIUS * sin(angle_i), 0.0f } });
+        builder.add_vertex({ { 0.5f * CIRCLE_COARSE_GRADE_IN_RADIUS * cosf(angle_i), 0.5f * CIRCLE_COARSE_GRADE_IN_RADIUS * sinf(angle_i), 0.0f } });
+        builder.add_vertex({ { 0.5f * CIRCLE_COARSE_GRADE_OUT_RADIUS * cosf(angle_i), 0.5f * CIRCLE_COARSE_GRADE_OUT_RADIUS * sinf(angle_i), 0.0f } });
     }
     builder
         .add_draw_command({ Render::PrimitiveType::Lines, 0, builder.vertex_count(),  Render::Material{}});
